backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
asyncpg==0.29.0
pgvector==0.2.5
pydantic==2.7.1
pdfplumber==0.11.0
spacy>=3.8.0,<3.9.0
sentence-transformers==2.7.0
numpy>=1.26.4
scikit-learn>=1.5.0
networkx==3.3
openai>=1.30.0
python-dotenv==1.0.1
